v0.3: remove duplicate dependency

+++ xpkg.py
@@ -6,8 +6,9 @@
 import os, shutil, json
 from pathlib import Path
 from argparse import ArgumentParser
+from collections import deque
 
-__version__ = "0.2"
+__version__ = "0.3"
 base_path = os.environ["SOFTWARE_BASE"]
 status_path = os.path.join(base_path, "var/xpkg-status.json")
 tmp_path = os.path.join(base_path, "tmp")
@@ -37,6 +38,7 @@ def parse_args():
         help="force install ignoring dependency errors",
     )
     parser.add_argument("--clear", action="store_true")
+    parser.add_argument("--update", action="store_true")
     parser.add_argument("--list", "-l", action="store_true")
     return parser.parse_args()
 
@@ -46,7 +48,7 @@ def fix_link():
         src = os.path.join(base_path, k)
         if not Path(src).is_symlink():
             dst = os.path.join(tmp_path, v)
-            shutil.copytree(src, dst)
+            shutil.copytree(src, dst, dirs_exist_ok=True)
             shutil.rmtree(src)
             os.symlink(dst, src)
 
@@ -79,19 +81,7 @@ def init():
     print(shell_config)
 
 
-def get_depends(name: str) -> list:
-    depends = (
-        os.popen("apt-cache depends {} | grep ' Depends'".format(name))
-        .read()
-        .splitlines()
-    )
-    return [x.split(": ")[1] for x in depends if not x.endswith(">")]
-
-
-def install_package(name: str, status: dict, force=False, manual=False) -> None:
-    if name in status:
-        print(name, "is already installed")
-        return
+def get_depends(name: str, manual=False) -> list[str]:
     if name.startswith("python3-"):
         if manual:
             print("Error:", name, "is a python3-only package, please use pip instead")
@@ -103,24 +93,48 @@ def install_package(name: str, status: dict, force=False, manual=False) -> None:
                 "is a python3-only package, using pip instead; note that this won't be recorded by xpkg",
             )
             os.system('python -m pip install "{}"'.format(name))
-            return
-    depends = get_depends(name)
+            return []
+    depends = (
+        os.popen("apt-cache depends {} | grep ' Depends'".format(name))
+        .read()
+        .splitlines()
+    )
+    depends = [x.split(": ")[1] for x in depends if not x.endswith(">")]
     if depends:
         print(name, "depends on ", depends, ", installing dependencies first")
-    for dep in depends:
-        install_package(dep, status, force)
-    if os.system("apt-get download " + name):
-        if force:
-            print("Error installing" + name + ": ignoring")
-            return
+    return depends
+
+
+def install_packages(names: list[str], status: dict, force=False, manual=False) -> None:
+    required = set()
+    processing = deque()
+    for name in names:
+        if name in status:
+            print(name, "is already installed")
         else:
-            raise RuntimeError("Error installing package " + name)
-    filename = [file for file in os.listdir(".") if file.endswith(".deb")][0]
-    files = (
-        os.popen("dpkg-deb -xv {} {}".format(filename, base_path)).read().splitlines()
-    )
-    status[name] = {"type": "xpkg", "files": files}
-    os.system("rm -f *.deb")
+            required.add(name)
+            processing.append(name)
+    while processing:
+        for name in get_depends(processing.popleft(), manual):
+            if name in status:
+                print(name, "is already installed")
+            elif name not in required:
+                required.add(name)
+                processing.append(name)
+    for name in required:
+        if os.system("apt-get download " + name):
+            if force:
+                print("Error installing " + name + ": ignoring")
+            else:
+                raise RuntimeError("Error installing package " + name)
+        filename = [file for file in os.listdir(".") if file.endswith(".deb")][0]
+        files = (
+            os.popen("dpkg-deb -xv {} {}".format(filename, base_path))
+            .read()
+            .splitlines()
+        )
+        status[name] = {"type": "xpkg", "files": files}
+        os.system("rm -f *.deb")
 
 
 def remove_package(name, status):
@@ -142,14 +156,19 @@ def remove_package(name, status):
 def main():
     if not os.path.exists(status_path):
         init()
+        return
     args = parse_args()
+    if args.update:
+        shutil.copy(__file__, Path(base_path) / "usr/sbin/xpkg")
+        os.system("chmod +x {}".format(Path(base_path) / "usr/sbin/xpkg"))
+        return
     if args.clear:
         if input("Are you sure to uninstall xpkg and clear all packages? [y/n]") == "y":
             shutil.rmtree(base_path)
             print("Bye!")
         else:
             print("Abort.")
-        exit(0)
+        return
     if not args.install and not args.remove and not args.list:
         print("type `xpkg --help' for help")
         exit(-1)
@@ -160,22 +179,22 @@ def main():
             "\n".join(key for key, value in status.items() if value["type"] == "xpkg")
         )
         exit(0)
-    for pkg in args.install:
-        if len([file for file in os.listdir(".") if file.endswith(".deb")]):
-            print(
-                "Error: current directory contains deb files. Please cd to another directory"
-            )
-            exit(-1)
-        install_package(pkg, status, args.force, True)
+    if len([file for file in os.listdir(".") if file.endswith(".deb")]):
+        print(
+            "Error: current directory contains deb files. Please cd to another directory"
+        )
+        exit(-1)
+    if args.install:
+        install_packages(args.install, status, args.force, True)
         fix_link()
         print(
-            "successfully installed package",
-            pkg,
+            "successfully installed packages ",
+            args.install,
             ", please source .bashrc or restart the shell to use",
         )
     for pkg in args.remove:
         remove_package(pkg, status)
-        print("successfully removed package", pkg)
+        print("successfully removed package ", pkg)
     with open(status_path, "w") as f:
         json.dump(status, f, indent=2)
